update
git-svn-id: http://svn.tudos.org/repos/oc/tudos/trunk@19 d050ee49-bd90-4346-b210-929a50b99cfc

+++ l4/pkg/valgrind/src/valgrind-3.6.0-svn/drd/tests/tsan_unittest.cpp
@@ -6620,8 +6620,8 @@ void Getter() {
     }
     usleep(1000);
   }
-  printf("T=%d: non_zero_received=%d\n", 
-         (int)pthread_self(), non_zero_received);
+  printf("T=%zd: non_zero_received=%d\n", 
+         (size_t)pthread_self(), non_zero_received);
 } 
 
 void Run() {
